Global systems, logging window
Implemented simple logging window - lacking functionality at the moment
but will be developed into a console-style system. Application and its
plugins will be able to print messages to the log window and all
messages will be logged to files. The logging window can be enabled
currently by passing "-debug" on the command line. Also implemented some
global pointers for convenience - g_pCmdLine holds a pointer to the
global command-line parser created in main.cpp before the event loop is
started, and g_pLog holds a pointer to the logging window. Most logging
functions such as showing the log window will not work if -debug is not
passed.