Extended logging system
Added functionality for logging to files. "-log" on the command line
enables logging to files only (as opposed to "-debug" which enables both
the log window and logging to files). Log files are named
[date]_[time].log. LogMessage(QString) logs a message to the window/file
as appropriate, and LogWarning(QString) logs a warning message (preceded
by "WARNING" in log file, bold and red in log window).