Added GLEW 1.9.0
Included GLEW 1.9.0 with the sources in order to avoid having to
hard-code include paths etc. in the .pro files. It appears Qt ships with
a generally rubbish OpenGL library - the example applications don't
compile and there are undeclared identifier errors everywhere, so an
external library is required for OpenGL support.