Fixed ConsoleWidget colour bug
That's a pain in the arse finally dealt with. The ConsoleWidget will now
update its content to reflect the new message colour whenever a colour
property is changed. The command entry box also keeps a list of commands
recently entered, which can be accessed with the arrow keys when the box
is empty.